pyproject.toml
[project]
name = "py-note"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyinstaller",
    "flask",
    "pywebview",
    "setuptools",
    "requests",
    "psutil",
    "pydantic",
    "pydantic-to-typescript",
]
